lab_7/app/backend/requirements.txt
fastapi==0.115.12
httpx==0.28.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-sdk==1.28.2
prometheus-client==0.21.0
psycopg[binary]==3.2.9
uvicorn[standard]==0.34.2
